refactor(tests): streamline meshtastic utility test suite
Merge multiple specialized test files into core utility suites to reduce redundancy and improve maintainability.

Changes:
- Integrated callback lifecycle and client cleanup tests into `test_meshtastic_utils_connect.py`.
- Consolidated reconnection logic and path tests into `test_meshtastic_utils_disconnect.py`.
- Moved message routing and path validation tests into `test_meshtastic_utils_messages.py`.
- Absorbed clock skew and packet drain coverage tests into primary connection suites.
- Expanded `reset_meshtastic_relay_state` fixture to ensure comprehensive global state isolation across consolidated tests.
