+++ tests/test_meshtastic_utils_disconnect.py
@@ -12,16 +12,20 @@ Tests the Meshtastic client functionality including:
 
 import asyncio
 import contextlib
+import sys
 import threading
+import types
 import unittest
 from concurrent.futures import TimeoutError as ConcurrentTimeoutError
 from typing import Any
-from unittest.mock import MagicMock, Mock, patch
+from unittest.mock import AsyncMock, MagicMock, Mock, patch
 
 import pytest
 
+import mmrelay.meshtastic_utils as mu
 from mmrelay.meshtastic_utils import (
     on_lost_meshtastic_connection,
+    reconnect,
 )
 from tests.conftest import cleanup_ble_future_state
 
@@ -106,6 +110,26 @@ def reset_meshtastic_relay_state(monkeypatch):
     monkeypatch.setattr(
         "mmrelay.meshtastic_utils._health_probe_request_deadlines",
         {},
+        raising=False,
+    )
+    monkeypatch.setattr(
+        "mmrelay.meshtastic_utils.reconnecting",
+        False,
+        raising=False,
+    )
+    monkeypatch.setattr(
+        "mmrelay.meshtastic_utils.shutting_down",
+        False,
+        raising=False,
+    )
+    monkeypatch.setattr(
+        "mmrelay.meshtastic_utils.meshtastic_client",
+        None,
+        raising=False,
+    )
+    monkeypatch.setattr(
+        "mmrelay.meshtastic_utils.reconnect_task_future",
+        None,
         raising=False,
     )
 
@@ -493,8 +517,6 @@ class TestConnectionLossHandling(unittest.TestCase):
         mmrelay.meshtastic_utils.shutting_down = False
 
         mock_interface = MagicMock()
-        mock_interface._last_disconnect_source = "   "  # Whitespace only
-
         on_lost_meshtastic_connection(
             mock_interface, detection_source="unknown", topic=pub.AUTO_TOPIC
         )
@@ -502,3 +524,276 @@ class TestConnectionLossHandling(unittest.TestCase):
         # Should fall back to default detection source
         error_call = mock_logger.error.call_args[0][0]
         self.assertIn("meshtastic.connection.lost", error_call)
+
+
+# ---------------------------------------------------------------------------
+# Helpers for reconnect tests (absorbed from reconnect + reconnect_paths)
+# ---------------------------------------------------------------------------
+
+
+def _sleep_and_mark_shutdown(_seconds: int) -> None:
+    mu.shutting_down = True
+
+
+class _DummyColumn:
+    def __init__(self, *args, **kwargs):
+        pass
+
+
+class _FailedExecutorLoop:
+    def __init__(self, future):
+        self._future = future
+
+    def run_in_executor(self, *_args, **_kwargs):
+        return self._future
+
+
+def _sleep_and_shutdown(_seconds):
+    mu.shutting_down = True
+    return None
+
+
+def _mark_shutdown(*_args, **_kwargs):
+    mu.shutting_down = True
+
+
+# ---------------------------------------------------------------------------
+# Reconnect tests absorbed from test_meshtastic_utils_reconnect.py
+# ---------------------------------------------------------------------------
+
+
+@pytest.mark.asyncio
+class TestReconnectSuccess:
+    async def test_reconnect_succeeds_and_clears_future_and_flag(self):
+        mock_client = MagicMock()
+        mu.reconnecting = True
+        mu.shutting_down = False
+        mu.reconnect_task_future = None
+
+        def _connect_side_effect(_cfg, _force):
+            mu.meshtastic_client = mock_client
+            return mock_client
+
+        with (
+            patch("mmrelay.meshtastic_utils.is_running_as_service", return_value=True),
+            patch(
+                "mmrelay.meshtastic_utils.connect_meshtastic",
+                side_effect=_connect_side_effect,
+            ),
+            patch("mmrelay.meshtastic_utils.asyncio.sleep", new_callable=AsyncMock),
+        ):
+            await reconnect()
+
+        assert mu.meshtastic_client is mock_client
+        assert mu.reconnecting is False
+        assert mu.reconnect_task_future is None
+
+    async def test_reconnect_success_does_not_republish_client_global(self):
+        mock_client = MagicMock()
+        mu.reconnecting = True
+        mu.shutting_down = False
+        mu.reconnect_task_future = None
+        mu.meshtastic_client = None
+
+        with (
+            patch("mmrelay.meshtastic_utils.is_running_as_service", return_value=True),
+            patch(
+                "mmrelay.meshtastic_utils.connect_meshtastic", return_value=mock_client
+            ),
+            patch("mmrelay.meshtastic_utils.asyncio.sleep", new_callable=AsyncMock),
+        ):
+            await reconnect()
+
+        assert mu.reconnect_task_future is None
+        assert mu.reconnecting is False
+        assert mu.meshtastic_client is None
+
+
+@pytest.mark.asyncio
+class TestReconnectCancellation:
+    async def test_reconnect_cancellation_logs_and_clears_state(self):
+        mu.reconnecting = True
+        mu.shutting_down = False
+        mu.reconnect_task_future = None
+
+        with (
+            patch("mmrelay.meshtastic_utils.is_running_as_service", return_value=True),
+            patch(
+                "mmrelay.meshtastic_utils.asyncio.sleep",
+                new_callable=AsyncMock,
+                side_effect=asyncio.CancelledError,
+            ),
+            patch("mmrelay.meshtastic_utils.logger") as mock_logger,
+        ):
+            await reconnect()
+
+        mock_logger.info.assert_any_call("Reconnection task was cancelled.")
+        assert mu.reconnecting is False
+        assert mu.reconnect_task_future is None
+
+
+@pytest.mark.asyncio
+class TestReconnectShutdownAbort:
+    async def test_shutdown_during_backoff_aborts_reconnect(self):
+        mu.reconnecting = True
+        mu.shutting_down = False
+        mu.reconnect_task_future = None
+
+        with (
+            patch("mmrelay.meshtastic_utils.is_running_as_service", return_value=True),
+            patch(
+                "mmrelay.meshtastic_utils.asyncio.sleep",
+                new_callable=AsyncMock,
+                side_effect=_sleep_and_mark_shutdown,
+            ),
+            patch("mmrelay.meshtastic_utils.connect_meshtastic") as mock_connect,
+            patch("mmrelay.meshtastic_utils.logger") as mock_logger,
+        ):
+            await reconnect()
+
+        mock_connect.assert_not_called()
+        mock_logger.debug.assert_any_call(
+            "Shutdown in progress. Aborting reconnection attempts."
+        )
+        assert mu.reconnecting is False
+
+
+@pytest.mark.asyncio
+class TestReconnectFailureBackoff:
+    async def test_connect_failure_logs_exception_and_clears_state(self):
+        mu.reconnecting = True
+        mu.shutting_down = False
+        mu.reconnect_task_future = None
+        attempt_count = 0
+
+        def _connect_side_effect(_cfg, _force):
+            nonlocal attempt_count
+            attempt_count += 1
+            if attempt_count >= 2:
+                mu.shutting_down = True
+            raise ConnectionError("connection refused")
+
+        with (
+            patch("mmrelay.meshtastic_utils.is_running_as_service", return_value=True),
+            patch(
+                "mmrelay.meshtastic_utils.connect_meshtastic",
+                side_effect=_connect_side_effect,
+            ),
+            patch("mmrelay.meshtastic_utils.asyncio.sleep", new_callable=AsyncMock),
+            patch("mmrelay.meshtastic_utils.logger") as mock_logger,
+        ):
+            await reconnect()
+
+        assert attempt_count == 2
+        assert any(
+            "Reconnection attempt failed" in str(c.args)
+            for c in mock_logger.exception.call_args_list
+        )
+        assert mu.reconnecting is False
+        assert mu.reconnect_task_future is None
+
+    async def test_reconnect_task_future_cleared_after_failure(self):
+        mu.reconnecting = True
+        mu.shutting_down = False
+        mu.reconnect_task_future = None
+        attempt_count = 0
+
+        def _connect_side_effect(_cfg, _force):
+            nonlocal attempt_count
+            attempt_count += 1
+            if attempt_count >= 2:
+                mu.shutting_down = True
+            raise RuntimeError("unexpected error")
+
+        with (
+            patch("mmrelay.meshtastic_utils.is_running_as_service", return_value=True),
+            patch(
+                "mmrelay.meshtastic_utils.connect_meshtastic",
+                side_effect=_connect_side_effect,
+            ),
+            patch("mmrelay.meshtastic_utils.asyncio.sleep", new_callable=AsyncMock),
+            patch("mmrelay.meshtastic_utils.logger"),
+        ):
+            await reconnect()
+
+        assert attempt_count == 2
+        assert mu.reconnect_task_future is None
+        assert mu.reconnecting is False
+
+
+# ---------------------------------------------------------------------------
+# Reconnect path tests absorbed from test_meshtastic_utils_reconnect_paths.py
+# ---------------------------------------------------------------------------
+
+
+@pytest.mark.asyncio
+async def test_reconnect_rich_progress_breaks_on_shutdown():
+    mock_progress_instance = MagicMock()
+    mock_progress_class = MagicMock()
+    mock_progress_class.return_value.__enter__.return_value = mock_progress_instance
+
+    fake_rich = types.ModuleType("rich")
+    fake_progress = types.ModuleType("rich.progress")
+    fake_progress.Progress = mock_progress_class
+    fake_progress.BarColumn = _DummyColumn
+    fake_progress.TextColumn = _DummyColumn
+    fake_progress.TimeRemainingColumn = _DummyColumn
+    fake_rich.progress = fake_progress
+
+    with (
+        patch.dict(sys.modules, {"rich": fake_rich, "rich.progress": fake_progress}),
+        patch("mmrelay.meshtastic_utils.DEFAULT_BACKOFF_TIME", 1),
+        patch("mmrelay.meshtastic_utils.is_running_as_service", return_value=False),
+        patch(
+            "mmrelay.meshtastic_utils.asyncio.sleep",
+            new_callable=AsyncMock,
+            side_effect=_sleep_and_shutdown,
+        ),
+        patch("mmrelay.meshtastic_utils.connect_meshtastic") as mock_connect,
+        patch("mmrelay.meshtastic_utils.logger") as mock_logger,
+    ):
+        await reconnect()
+
+    mock_connect.assert_not_called()
+    mock_progress_instance.update.assert_called_once()
+    mock_logger.debug.assert_any_call(
+        "Shutdown in progress. Aborting reconnection attempts."
+    )
+
+
+@pytest.mark.asyncio
+async def test_reconnect_logs_exception_and_backs_off():
+    running_loop = asyncio.get_running_loop()
+    failed_future = running_loop.create_future()
+    failed_future.set_exception(RuntimeError("boom"))
+    loop = _FailedExecutorLoop(failed_future)
+
+    with (
+        patch("mmrelay.meshtastic_utils.is_running_as_service", return_value=True),
+        patch(
+            "mmrelay.meshtastic_utils.asyncio.get_running_loop",
+            return_value=loop,
+        ),
+        patch("mmrelay.meshtastic_utils.asyncio.sleep", new_callable=AsyncMock),
+        patch("mmrelay.meshtastic_utils.logger") as mock_logger,
+    ):
+        mock_logger.exception.side_effect = _mark_shutdown
+        await reconnect()
+
+    mock_logger.exception.assert_called_once()
+
+
+@pytest.mark.asyncio
+async def test_reconnect_logs_cancelled():
+    with (
+        patch("mmrelay.meshtastic_utils.is_running_as_service", return_value=True),
+        patch(
+            "mmrelay.meshtastic_utils.asyncio.sleep",
+            new_callable=AsyncMock,
+            side_effect=asyncio.CancelledError,
+        ),
+        patch("mmrelay.meshtastic_utils.logger") as mock_logger,
+    ):
+        await reconnect()
+
+    mock_logger.info.assert_any_call("Reconnection task was cancelled.")
